+++ lib/state_machine/event.rb
@@ -193,14 +193,16 @@ module StateMachine
     # create 1 or more edges on the graph for each branch (i.e. transition)
     # configured.
     # 
-    # A collection of the generated edges will be returned.
-    # 
     # Configuration options:
     # * <tt>:human_name</tt> - Whether to use the event's human name for the
     #   node's label that gets drawn on the graph
     def draw(graph, options = {})
       valid_states = machine.states.by_priority.map {|state| state.name}
-      branches.collect {|branch| branch.draw(graph, options[:human_name] ? human_name : name, valid_states)}.flatten
+      branches.each do |branch|
+        branch.draw(graph, options[:human_name] ? human_name : name, valid_states)
+      end
+      
+      true
     end
     
     # Generates a nicely formatted description of this event's contents.